Split bip conf into different folders
