+++ bip_utils/bip/bip44/bip44.py
@@ -22,7 +22,7 @@
 # Imports
 from typing import Dict
 from bip_utils.bip.bip32 import Bip32Utils
-from bip_utils.bip.bip44_base.bip44_base import Bip44Base, Bip44Changes
+from bip_utils.bip.bip44_base import Bip44Changes, Bip44Base
 from bip_utils.conf import Bip44Coins, BipCoinConf, Bip44ConfGetter
