+++ bip_utils/bip/bip44_base/__init__.py
@@ -1,3 +1,3 @@
 from bip_utils.bip.bip44_base.bip44_base_ex import Bip44DepthError, Bip44CoinNotAllowedError
-from bip_utils.bip.bip44_base.bip44_base import Bip44Changes, Bip44Coins, Bip44Levels
+from bip_utils.bip.bip44_base.bip44_base import Bip44Changes, Bip44Levels, Bip44Base
 from bip_utils.bip.bip44_base.bip44_keys import Bip44PublicKey, Bip44PrivateKey
