+++ bip_utils/conf/__init__.py
@@ -1,10 +1,4 @@
-from bip_utils.conf.bip_coin_conf_helper import CoinNames, KeyNetVersions
-from bip_utils.conf.bip_coin_conf_enum import AddrTypes, Bip32Types
-from bip_utils.conf.bip_coin_conf import BipCoinConf
-from bip_utils.conf.bip44_coins import Bip44Coins
-from bip_utils.conf.bip44_coin_conf import *
-from bip_utils.conf.bip49_coin_conf import *
-from bip_utils.conf.bip84_coin_conf import *
-from bip_utils.conf.bip44_conf_getter import Bip44ConfGetter
-from bip_utils.conf.bip49_conf_getter import Bip49ConfGetter
-from bip_utils.conf.bip84_conf_getter import Bip84ConfGetter
+from bip_utils.conf.common import *
+from bip_utils.conf.bip44 import *
+from bip_utils.conf.bip49 import *
+from bip_utils.conf.bip84 import *
