+++ bip_utils/conf/bip44/__init__.py
@@ -1,0 +1,3 @@
+from bip_utils.conf.bip44.bip44_coins import Bip44Coins
+from bip_utils.conf.bip44.bip44_coin_conf import *
+from bip_utils.conf.bip44.bip44_conf_getter import Bip44ConfGetter
