+++ bip_utils/conf/bip44/bip44_coins.py
