+++ bip_utils/conf/bip44/bip44_conf_getter.py
@@ -21,9 +21,9 @@
 
 # Imports
 from typing import Dict
-from bip_utils.conf.bip_coin_conf import BipCoinConf
-from bip_utils.conf.bip44_coins import Bip44Coins
-from bip_utils.conf.bip44_coin_conf import *
+from bip_utils.conf.common import BipCoinConf
+from bip_utils.conf.bip44.bip44_coins import Bip44Coins
+from bip_utils.conf.bip44.bip44_coin_conf import *
 
 
 class Bip44ConfGetterConst:
