+++ bip_utils/conf/bip84/bip84_coin_conf.py
@@ -20,10 +20,7 @@
 
 
 # Imports
-from bip_utils.conf.bip_coin_conf_enum import AddrTypes, Bip32Types
-from bip_utils.conf.bip_coin_conf_common import *
-from bip_utils.conf.bip_coin_conf_helper import CoinNames, KeyNetVersions
-from bip_utils.conf.bip_coin_conf import BipCoinConf
+from bip_utils.conf.common import *
 
 
 # Bitcoin key net version (zpub / zprv)
