+++ bip_utils/conf/common/bip_coin_conf_const.py
